pyproject.toml
[tool.poetry]
name = "scraping"
version = "0.1.0"
description = ""
authors = ["arthurlemon <coucou@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
selenium = "^4.26.1"
beautifulsoup4 = "^4.12.3"
webdriver-manager = "^4.0.2"
selectolax = "^0.3.26"
pydantic = "^2.10.3"
playwright = "^1.49.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.notebook.dependencies]
jupyterlab = "^4.3.1"
jupytext = "^1.16.4"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
